feat: Implement initial Delivery module for order processing, including API, mock service integrations, and comprehensive tests.

+++ database/migrations/2026_02_12_104812_create_order_logs_table.php
@@ -18,7 +18,7 @@ return new class extends Migration {
             $table->json('request_body')->nullable();
             $table->json('response_body')->nullable();
             $table->integer('status_code');
-            $table->integer('duration_ms');
+            $table->float('duration_ms');
             $table->timestamps();
         });
     }
